+++ src/lib/libast/sfio/sfdlen.c
@@ -2,7 +2,7 @@
 *                                                                      *
 *               This software is part of the ast package               *
 *          Copyright (c) 1985-2011 AT&T Intellectual Property          *
-*          Copyright (c) 2020-2024 Contributors to ksh 93u+m           *
+*          Copyright (c) 2020-2026 Contributors to ksh 93u+m           *
 *                      and is licensed under the                       *
 *                 Eclipse Public License, Version 2.0                  *
 *                                                                      *
@@ -14,6 +14,7 @@
 *                  David Korn <dgk@research.att.com>                   *
 *                   Phong Vo <kpv@research.att.com>                    *
 *                  Martijn Dekker <martijn@inlv.org>                   *
+*            Johnothan King <johnothanking@protonmail.com>             *
 *                                                                      *
 ***********************************************************************/
 #include	"sfhdr.h"
@@ -25,7 +26,7 @@
 
 int _sfdlen(Sfdouble_t v)
 {
-#define N_ARRAY		(16*sizeof(Sfdouble_t))
+#define N_ARRAY		((int)(16*sizeof(Sfdouble_t)))
 	int		n, w;
 	Sfdouble_t	x;
 	int		exp;
