+++ src/lib/libast/stdio/snprintf.c
@@ -2,7 +2,7 @@
 *                                                                      *
 *               This software is part of the ast package               *
 *          Copyright (c) 1985-2011 AT&T Intellectual Property          *
-*          Copyright (c) 2020-2022 Contributors to ksh 93u+m           *
+*          Copyright (c) 2020-2026 Contributors to ksh 93u+m           *
 *                      and is licensed under the                       *
 *                 Eclipse Public License, Version 2.0                  *
 *                                                                      *
@@ -14,19 +14,20 @@
 *                  David Korn <dgk@research.att.com>                   *
 *                   Phong Vo <kpv@research.att.com>                    *
 *                  Martijn Dekker <martijn@inlv.org>                   *
+*            Johnothan King <johnothanking@protonmail.com>             *
 *                                                                      *
 ***********************************************************************/
 
 #include "stdhdr.h"
 
 int
-snprintf(char* s, int n, const char* fmt, ...)
+snprintf(char* s, size_t n, const char* fmt, ...)
 {
 	va_list	args;
 	int	v;
 
 	va_start(args, fmt);
-	v = sfvsprintf(s, n, fmt, args);
+	v = (int)sfvsprintf(s, n, fmt, args);
 	va_end(args);
 	return v;
 }
